as you requested, heath percent is a float

+++ player.py
@@ -32,7 +32,7 @@ class Player():
             self.tempH += val
 
     def getHealthPercent(self):
-        return int((float(self.currentH) / self.maxH) * 100)
+        return (float(self.currentH) / self.maxH) * 100
 
     def __str__(self):
         return "{} / {}".format(self.currentH, self.maxH)
